requirements.txt
PyGithub[integrations]
